refactor/#142: lambda -> lambda.recruit 패키지 이동
