+++ src/main/java/com/itstime/xpact/domain/recruit/service/DetailRecruitService.java
@@ -4,7 +4,7 @@ import com.itstime.xpact.domain.recruit.entity.DetailRecruit;
 import com.itstime.xpact.domain.recruit.entity.Recruit;
 import com.itstime.xpact.domain.recruit.repository.DetailRecruitRepository;
 import com.itstime.xpact.domain.recruit.repository.RecruitRepository;
-import com.itstime.xpact.infra.lambda.LambdaUtil;
+import com.itstime.xpact.infra.lambda.recruit.LambdaUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
