+++ src/main/java/com/itstime/xpact/domain/recruit/service/RecruitService.java
@@ -9,7 +9,7 @@ import com.itstime.xpact.domain.recruit.repository.DetailRecruitRepository;
 import com.itstime.xpact.domain.recruit.repository.RecruitRepository;
 import com.itstime.xpact.global.auth.SecurityProvider;
 import com.itstime.xpact.global.exception.CustomException;
-import com.itstime.xpact.infra.lambda.LambdaUtil;
+import com.itstime.xpact.infra.lambda.recruit.LambdaUtil;
 import com.itstime.xpact.global.exception.GeneralException;
 import com.itstime.xpact.global.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
