+++ src/main/java/com/itstime/xpact/infra/lambda/recruit/LambdaUtil.java
@@ -1,4 +1,4 @@
-package com.itstime.xpact.infra.lambda;
+package com.itstime.xpact.infra.lambda.recruit;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
